chmod calculator in octal.
Numerical permissions
The chmod numerical format accepts up to four digits. The three rightmost digits define permissions for the file user, the group, and others. The optional leading digit, when 4 digits are given, specifies the special setuid, setgid, and sticky flags. Each digit of the three rightmost digits represents a binary value, which controls the "read", "write" and "execute" permissions respectively. A value of 1 means a class is allowed that action, while a 0 means it is disallowed.

#	Sum	rwx	Permission
7	4(r) + 2(w) + 1(x)	rwx	read, write and execute
6	4(r) + 2(w)	rw-	read and write
5	4(r)        + 1(x)	r-x	read and execute
4	4(r)	r--	read only
3	       2(w) + 1(x)	-wx	write and execute
2	       2(w)	-w-	write only
1	              1(x)	--x	execute only
0	0	---	none
For example, 754 would allow:

"read" (4), "write" (2), and "execute" (1) for the User class i.e. 7 (4+2+1).
"read" (4) and "execute" (1) for the Group class i.e. 5 (4+1).
Only "read" (4) for the Others class.
A numerical code permits execution if and only if it is odd (i.e. 1, 3, 5, or 7). A numerical code permits "read" if and only if it is greater than or equal to 4 (i.e. 4, 5, 6, or 7). A numerical code permits "write" if and only if it is 2, 3, 6, or 7.